% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{cspServer}
\alias{cspServer}
\title{cspServer}
\usage{
cspServer(input, output)
}
\arguments{
\item{input, }{input params from UI}

\item{output, }{output params to UI}
}
\value{
the panel for main plots;
}
\description{
Sets up shinyServer to be able to run crisprseekplus interactively.
}
\note{
\code{cspServer}
}
\examples{
    cspServer

}

