% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cspUI}
\alias{cspUI}
\title{cspUI}
\usage{
cspUI()
}
\value{
the panel for the UI;
}
\description{
Creates a shinyUI to be able to run crsiprseekplus interactively.
}
\note{
\code{cspUI}
}
\examples{
    x<-cspUI()

}

