% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disableDownload.R
\name{disableDownload}
\alias{disableDownload}
\title{disableDownload}
\usage{
disableDownload(input)
}
\arguments{
\item{input, }{disable the download button}
}
\value{
the download button either enabled or disabled;
}
\description{
Enable or disable to download button depending on if analysis is complete
}
\note{
\code{disableDownload}
}
\examples{
    x<- disableDownload()

}

