% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInputFunc.R
\name{fileInputFunc}
\alias{fileInputFunc}
\title{fileInputFunc}
\usage{
fileInputFunc(input, sampleFile)
}
\arguments{
\item{input, }{enter correct files}

\item{sampleFile, }{sampleFile}
}
\value{
If no file is uploaded, use sample file
}
\description{
If input file is empty, use sample file
}
\note{
\code{fileInputFunc}
}
\examples{
    x<- fileInputFunc()


}

