% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startcrisprseekplus}
\alias{startcrisprseekplus}
\title{startcrisprseekplus}
\usage{
startcrisprseekplus()
}
\value{
the app
}
\description{
Starts the crisprseekplus to be able to run.
}
\note{
\code{startcrisprseekplus}
}
\examples{
    startcrisprseekplus()

}

