% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueFalseFunc.R
\name{trueFalseFunc}
\alias{trueFalseFunc}
\title{trueFalseFunc}
\usage{
trueFalseFunc(input)
}
\arguments{
\item{input, }{user inputs}
}
\value{
true or false depending on input
}
\description{
If radio button input == 1, the function returns true
and if the radio button value is 2, the function returns false
}
\note{
\code{trueFalseFunc}
}
\examples{
    x<- trueFalseFunc()


}

